+++ frontend/javascripts/dashboard/dataset/dataset_import_view.js
@@ -131,6 +131,7 @@ class DatasetImportView extends React.PureComponent<Props, State> {
         color: [255, 255, 255],
         intensityRange: [0, 255],
         isDisabled: false,
+        isInverted: false,
       };
       const datasetDefaultConfiguration = (await getDatasetDefaultConfiguration(
         this.props.datasetId,